+++ src/test/java/org/testcontainers/containers/CartridgeContainerTestUtils.java
@@ -1,12 +1,12 @@
 package org.testcontainers.containers;
 
-import java.util.Arrays;
-import java.util.List;
+import java.util.ArrayList;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 /**
  * @author Vladimir Rogach
+ * @author Ivan Dneprov
  */
 public class CartridgeContainerTestUtils {
 
@@ -15,10 +15,10 @@ public class CartridgeContainerTestUtils {
 
     static public void executeProfileReplaceSmokeTest(TarantoolCartridgeContainer container) throws Exception {
         container.executeCommand(
-                "return profile_replace(...)", Arrays.asList(1, "Ivanov Ivan Ivanovich", 33, 100500)).get();
+                "return profile_replace({1, \"Ivanov Ivan Ivanovich\", 33, 100500})");
 
-        List<?> result = container.executeCommand("return profile_get(...)", 1).get();
+        ArrayList<?> result = container.executeCommandDecoded("return profile_get(1)");
         assertEquals(1, result.size());
-        assertEquals(33, ((List<?>) result.get(0)).get(3));
+        assertEquals(33, ((ArrayList<?>) result.get(0)).get(3));
     }
 }