+++ src/test/java/org/testcontainers/containers/TarantoolContainerTest.java
@@ -2,12 +2,13 @@ package org.testcontainers.containers;
 
 import org.junit.jupiter.api.Test;
 
-import java.util.List;
+import java.util.ArrayList;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 /**
  * @author Alexey Kuzin
+ * @author Ivan Dneprov
  */
 class TarantoolContainerTest {
 
@@ -15,9 +16,8 @@ class TarantoolContainerTest {
     public void testExecuteScript() throws Exception {
         try (TarantoolContainer container = new TarantoolContainer()) {
             container.start();
-
-            container.executeScript("org/testcontainers/containers/test.lua").get();
-            List<?> result = container.executeCommand("return user_function_no_param()").get();
+            container.executeScript("org/testcontainers/containers/test.lua");
+            ArrayList<?> result = container.executeCommandDecoded("return user_function_no_param()");
             assertEquals(1, result.size());
             assertEquals(5, result.get(0));
         }
@@ -35,15 +35,15 @@ class TarantoolContainerTest {
                 .withLogLevel(TarantoolLogLevel.INFO)) {
             container.start();
 
-            List<?> result = container.executeCommand("return box.cfg.memtx_memory").get();
+            ArrayList<?> result = container.executeCommandDecoded("return box.cfg.memtx_memory");
             assertEquals(1, result.size());
             assertEquals(memory, result.get(0));
 
-            result = container.executeCommand("return box.cfg.log_level").get();
+            result = container.executeCommandDecoded("return box.cfg.log_level");
             assertEquals(1, result.size());
             assertEquals(5, result.get(0));
 
-            result = container.executeCommand("return user_function_no_param()").get();
+            result = container.executeCommandDecoded("return user_function_no_param()");
             assertEquals(result.size(), 1);
             assertEquals(result.get(0), 5);
         }